Create Full Changes List
A full list of the changes made of this mod as of vRimworld 1.0

It may not make much sense to you if you don't understand much about the thought defs in the game or the xml that makes up the moods and situations.

But you can probably infer a bit of information from it anyway. 
I hope it helps those seeking to better understand the full workings of the mod.

Thank you to Rainbeau Flambe for his work compiling this list. He is awesome.